Add tags to plays
Adds various tags to the deploy_steps_playbook output.

- facts: Run fact gathering
- overcloud: Run all plays for overcloud deployment
- pre_deploy_steps: Run deployments that happen pre deploy_steps
- host_prep_steps: Run host_prep_tasks
- deploy_steps: Run deploy_steps
- post_deploy_steps: Run deployments that happen post deploy_steps
- external: Run all external deployments
- external_deploy_steps: Run all external deployments

external and external_deploy_steps are the same for now. However, I kept
them both b/c "external" is the parallel for "overcloud", and
"external_deploy_steps" is the parallel for "deploy_steps".

Also fixes the name of the host_prep_tasks play that was incorrectly
named "Deployment steps".

Change-Id: Ic6579d187035d00edc1f9190f8ebd12d4400d919