Pass the CloudDomain parameter down to hiera on all nodes
We currently do not have a single hiera key that simply
returns the CloudDomain (Domain Suffix) that was configured
during the deployment. As it is a rather useful thing to have
on the overcloud nodes, let's pass it down for all roles.

IHA is at least one user where this information is needed in
hiera.

Change-Id: I7a510e3bbeb367a46bf6ccc565393b1e7a1a1f33