Polish swift containers logging configuration
The logging directories are not used by swift, which in turn logs
to syslog. So the unnecessary bind-mounts were removed.

Furthermore, as part of the swift package which is included in the
overcloud image, there is an rsyslog rule which gets the swift logs
and persists them in /var/log/swift. Due to this and to be consistent
with the rest of the containers. A symlink was created to have the
swift logs available from /var/log/containers/swift as well. This makes
the readme indicating the change of place of the logs unnecessary, so
that was removed as well.

Closes-Bug: #1732107
Change-Id: I61610d3cb187235ac316e1c5de0a344be3ebb1e2