Run rabbitmq_policy,rabbitmq_user with pacemaker
When using the docker/services/pacemaker/rabbitmq.yaml service
it appears that custom rabbit users and policies aren't being
initialized.

For non-pacemaker containerized rabbit this typically occurs
in the docker_puppet_tasks. With pacemaker enabled however we
are doing this sort of initialization via the _init_bundle
container.

This causes issues for some integrations (Veritas) that require
extra rabbit users.

Change-Id: I521c7468ef3677ebfd645f609c0b76ba30df7c86