Remove too opinionated parameters from kubernetes.yaml environment
The settings being removed are remnants from early days of testing.

The flavor settings can directly conflict/confuse deployments with
quickstart, as the flavors are called oooq_control and oooq_compute
instead, and we shouldn't really need to set them here.

The ComputeServices setting will start breaking deployments as soon as
we want to deploy any APBs on the compute nodes.

Change-Id: I68d958a408594b169bd9bb09607f04c180f915b9