Update Iscsid puppet_config to reset host's IQN
The Iscsid service uses puppet to update the IQN of each node, so that
all overcloud nodes have a unique IQN. When Iscsid is containerized,
the new IQN is captured in the "puppet-generated" directory, and made
available to other containers using bind mounts.

However, any non-containerized service that uses the IQN will be seeing
the host's value, instead of the new value made available to the
containers. This IQN mismatch between containerized and non-containerized
services will cause failures, typically with authentication.

The problem is resolved by adding a bind mount to Iscsid's puppet_config
that allows the iscsid container to update the host's IQN.

Change-Id: I1cbcb858754cc4792a085438eb28f2476245228c