Swap the order of stdout and stderr in debug output
Generally this data is looked at because something failed, and in
that case the relevant error is likely to be at the end of stderr.
By concatenating the output stderr first and then stdout as we were
it is possible for the stderr to get lost entirely in the failures
list, and even if that doesn't happen it's best to output the
relevant error right at the end of the output where people will
see it.  Previously it would be buried in the middle of the debug
output.

Closes-bug: #1743754

Change-Id: I952fd1af5778ade1eb6b0599d983f98cadeb7f6f
(cherry picked from commit 5595e7fc14430ee0636060afc82f909d4b3d04be)