Ensure os-net-config conditional for upgrade doesn’t fail.
When using grep the check will fail if no upgrade is available.  This
has the side effect of not registering the variable and the check
later on will fail as “.stdout” property is not defined.

Using awk doesn’t have this side effect and will return empty stdout
without error in case of error.

Change-Id: I9d9bb3b625c2f0fb5f4a2c98bd3bf903f4bec176
Closes-Bug: #1733924