Fix neutron agent start order
In the baremetal deployment, we used to ensure that neutron-server was
started prior to starting up the various agents. In the containerized
deployment we need to ensure that we launch the agents after the server
has been started. We can do this by configuring a start_order for each
of the services.

It should be noted that the ovs agent was actually configured to start
in step5 on baremetal due to previous race conditions under HA
deployments. This change leaves it in step4 but configures the
start_order to be after the neutron-api service.

Change-Id: I3794400ef5c8ae620961914831ff85e3438b0399
Closes-Bug: #1734976
Related-Bug: #1663273
(cherry picked from commit c54c6222d8b517b8ee492caddaf47d2ef5780f31)