Do not use lsns to kill non-containerized epmd on the host
In a containerized HA environment, the Erlang port mapper daemon is
started via pacemaker. If an epmd runs on the host, it must be killed
to prevent it from holding the epmd port.

The current way of searching for non-containerized epmd is broken in
case lsns returns an error. Use pgrep's namespace filtering instead.

Change-Id: I29d451716c8e955b9e114a289ac1e22d2af35f1a
Closes-Bug: #1735698