zuul: remove file rule for ovb-ha
It was a bad idea because now we run ovb-ha only when we touch this
file. We don't want that. In fact, we have a irrelevant-files list in
tripleo-ci that we should improve if we know some files that have no
impact on ovb-ha.

This patch remove the file rule, so we run the job again on THT unless a
patch only touch irrelevant-files.

Change-Id: I7a2f8791bf52c2251bd8e13100fd2d50632008fc