undercloud: set default_resource_class to 'baremetal'
We already default to this value when enrolling nodes with our workflows, and when
upgrading an undercloud to Pike. This change ensures that nodes enrolled outside
of our workflows also default to 'baremetal'.

A similar change to instack-undercloud is Iee8604a6a942c12e6fa6b9cb8ea88a2cee1ead8f

Change-Id: I8c3c45810571a5d58d8d3c21561b96186d2ebbb2
Partial-Bug: #1732190