Use StrictHostKeyChecking=no in enable-ssh-admin.sh
Use StrictHostKeyChecking=no to inject the temporary ssh key in
enable-ssh-admin.sh. The user provides the list of hosts for ssh, so we
can safely assume that they intend to ssh to those hosts.

Also, for the ovb case the hosts will have new host ssh keys which have
not yet been accepted.

Change-Id: Ia2a2d1e4c214d9d9190131c280a54eab066e3657
Closes-Bug: #1737027