Wait for rabbitmq_ready tag
We need to wait for rabbitmq_ready exec so that rabbit is fully
up. This can only happen if we add the tag for it.
Also we need to make sure that launching the epmd process cannot
happen. The reason for this is the following:
When the puppet-rabbitmq module gets invoked (a simple facter run
will be sufficient) inside the rabbitmq_init_bundle container it spawns
an epmd process.
Now if we wait for the Exec[rabbitmq-ready], it means that this epmd
process is staying around until rabbit is up, but then will disappear
suddenly when the rabbitmq_init_bundle container exits, which will
subsequently confuse the rabbitmq cluster and make it fail.

Partial-Bug: #1739026

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Co-Authored-By: John Eckersberg <jeckersb@redhat.com>

Change-Id: Ie74a13a6c8181948900ea0de8ee9717f76f3ce79
(cherry picked from commit bab6ec25323fd9a667efc08825e4bc7b1732f056)