Add proper debug switch on init_bundles
When deploying with -e environments/config-debug.yaml, which sets
ConfigDebug to true, it is expected that puppet is run with --debug
--verbose. This has happened for most of the puppet uses (via
LP#1722752), but we missed enabling it for the init_bundle under
docker/services.

While we're at it we also add '--color=false' to the puppet apply
command of the init_bundle containers as that is what we use in the
other puppet apply runs.

NB: Cherry-pick not 100% clean on docker/services/haproxy.yaml due to
    the logging to stdou/stderr changes

Closes-Bug: #1738764

Change-Id: If529b83a7342b3ad17d705517978539d1c6b949e
(cherry picked from commit 4d7e03be85ce8a005aceca9cd1784b7cc5ad99eb)