Configure ODL Logging mechanism
ODL logs to either console for containarised deployment and to file
for non-containarised deployment. For containarised deployments, logs
can then be read via "docker logs".

We need both JVM and karaf (ODL application) logs together to debug
any failure scenario and for them to correlate. JVM logs to console
and not to a file. Karaf can log to file and console. So we make
karaf and JVM both log to console and read these logs via docker logs.

In cases when ODL container restarts, "docker logs" retain logs from
previous broken container so there is no loss of logs in these cases.

Change-Id: I2fe56df082c5d9206015f156f5f1b3cfca63c982