Add MinPoll and MaxPoll options to NTP module
These options specify the minimum and maximum poll intervals
for NTP messages, in seconds to the power of two.

The maximum poll interval defaults to 10 (1,024 s), but can be
increased by the MaxPoll option to an upper limit of 17 (36.4 h).
The minimum poll interval defaults to 6 (64 s), but can be decreased
by the MinPoll option to a lower limit of 4 (16 s).

Change-Id: Ib2929be86e8cb31c00d166abe750354188302415
Closes-bug: #1736170