Use relative path in ipv6 nic configs
The path to run-os-net-config.sh in these nic config templates for ci
should be relative, not absolute. The absolute path does not honor a
different directory specified via --templates, and it also fails the
overcloud deploy entirely b/c the absolute path does not exist in the
plan in swift.

Change-Id: I85ac081571965f42beef4bd283db985b9a03078f