Introduce tests for yaql structures
We use yaql in many hot templates, but they are usually faily obscure
and not documented, which makes it hard to change them over time (to
make them perform better for example).

This introduces some basic infrastructure for testing them, as well as a
first modification to show how things would be easier to change.

Change-Id: I1a8d3a514d8fa0de4b024d3ae320723d791f3e7d