Undercloud: support for external VIP for SSL
Wires in the use of ../network/ports/external_from_pool.yaml
so that we can control the named 'public_virtual_ip' created in
overcloud.yaml when using Undercloud SSL.

By default this has previously gone to the ctlplane IP for the
Undercloud itself. When SSL is enabled we want it set to a different
VIP managed by Haproxy/keepalived. If SSL is disabled
python-tripleoclient just sets it to the ctlplane so the previous
behavior is preserved.

Change-Id: Id8127efc658f4bae3176d7394a32face6030303c