Do not format output for kubespray or openshift deployments
The output comes from ansible and is already fully readable as it is.
Also, because the previous task didn't have the 'failed_when: false'
directive, it would never reach the 'print xxx outputs' task in case of
failure, while showing the output twice on success.

It is safe to just delete the task.

Change-Id: I56b44aec0a549e184f46344ea362f655ab80b3b0