Add a ComputePPC64LE role
Create a new role for compute nodes running on ppc64le.  Ideally we
wouldn't need this but at least the overcloud image needs to differ
based on CPU architecture so having a separate role allows us to vary
other parameters.

As the intent is for the same services to run on moth Compute and
ComputePPC64LE add a validation check to ensure that the services stay
in sync.

Blueprint: multiarch-support
Change-Id: I9912139048fa11df447c0985e7d53fbcfee7e252
(cherry picked from commit 40282174ea632e111c021507a3aaa1e7452aaabc)