Fixes SSL/TLS with OpenDaylight docker service
The ODL private key, cert, and CA cert were missing in the puppet
container config.  These are required during puppet stage in order to
create the proper keystores in ODL.  The files are not needed during the
service bringup time, because the information is built into the
keystore created by puppet.

Closes-Bug: 1747700

Change-Id: If548fdba836104412bf20e8e05ecf6a5058aa318
Signed-off-by: Tim Rozet <trozet@redhat.com>