Restrict SNMP to internal network
Add a parameter, SnmpdIpSubnet, which can be an IP/MASK that will be
used to secure with IPtables the source network authorized to reach
SNMP service on the host.
If SnmpdIpSubnet is left empty (default) the parameter will be set to
SnmpdNetwork.

Also change the IPtables id, 127 was used by Horizon, so let's switch
SNMP to 124. No impact on users.

Note: in this backport we also change the heat_template_version for SNMP
service so we support "conditions".

Change-Id: I46fce28926cb5a881f7384948480266712ae75e3
Depends-On: Ib203161b9676dcfaaf46eec2bddf767ec49282f7
Closes-Bug: #1749324
(cherry picked from commit 43155ed1462a8e27c9efdbb345bfc5832c50bd2f)