Add non-production ceph defaults to low-memory-usage.yaml
Ceph Luminous does not create a pool if the pg_number,
pool size, and mon_max_pg_per_osd are outside of Ceph
recomended practice for production clusters. TripleO
development environments which use low-memory-usage.yaml
may not meet this criteria and fail a deployment with
Luminous unless the defaults for these values are overriden
as in this change.

Change-Id: I12ee495b780f29fc098c5c3bd57c46fd946146ae
Closes-Bug: #1749544