Align zaqar max_messages_post_size with undercloud
When deploying a containerized undercloud, we need the same
configuration as we used to have in instack-undercloud.

For max_messages_post_size, we used to have 1048576 due to the high size
of the messages with config-download.

Let's sync the config here, so we can deploy an overcloud with
config-download when a containerized undercloud is used.

Change-Id: Ib43c811d9ea4e71558c15c78cfb9999f738b8098