Disable UseDNS in sshd config
The UseDNS causes the sshd server to perform a reverse DNS lookup when
a client connects which introduces reasonable delay when everything is
fine, but can be counted in seconds when DNS is not setup correctly.
This is particularly annoying when using ansible to configure the node
as it is failing with timeouts waiting for privilege escalation prompt.

This commit disables the UseDNS option since we do not use hostname
based authentication in TripleO.

Change-Id: Ib000f3488326a802d97d0bcdb3299d54248f0935