mistral-executor: mount /var/lib/mistral
When containerizing mistral-executor, we need to mount /var/lib/mistral so
our operators can get the config-download logs when the undercloud is
containerized and config-download is used to deploy the overcloud.

To help our operators, we also create /var/lib/mistral/readme.txt so
they know where to find the config-download data.

Change-Id: I8d31d5fec2721c6e4f82b1ad2169a7635cb57600
Closes-Bug: #1749823