Upgrade data on disk on mariadb major upgrade
Add an ansible task to run mysql_upgrade whenever a container
image upgrade causes a major upgrade of mariadb (e.g. 5.5 -> 10.1)

. If the overcloud was containerized prior to the major upgrade, the
  mysql upgrade job is ran in an ephemeral container (where the latest
  version of mysql comes from) and uses credentials from the Kolla
  configuration.

. Otherwise the upgrade job is run from the host (once the mysql
  rpm has been updated) and uses credentials from the host.

We log the output of the script in the journal. Also, the mysql server
needs to be started temporarily, so use a temporary log file for it
when run from the ephemeral container.

Change-Id: Id330d634ee214923407ea893fdf7a189dc477e5c
(cherry picked from commit 624fedb114d75d63ec0082948820eb380abc6a39)