Handle NodeTLSData resources for controller roles
In ocata (and newton), a custom role in place of the default controller
didn't work, b/c not all of the controller specific resources were
represented in role.role.j2.yaml.

This patch makes it so that at least overcloud SSL configuration is
handled by checking to see if the string "controller" is in the role
name. This fixes the issue of using deployed-server with ssl for ocata.

Change-Id: I7e8487208cc71a0780abc167d06902b9f3d3c983
Closes-Bug: #1751179