Fixes OpenDaylight healthcheck/GUI feature
Deployments with OpenDaylight fail because the Docker healthcheck fails.
This is due to the fact that the URL used for querying with the
healthcheck is not available, due to a feature mismatch.  This patch
fixes the GUI feature being installed to address the issue.

Closes-Bug: 1751857

Change-Id: Ia62aebb9c03e09c1986c1592a38ab455e68fd103
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit e581c27e32202606e4580b6ba8b42c3bb9ae097e)