Disable murano horizon plugin until dep is met
Murano dashboard has a dependency on heat-dashboard which is not yet in
the horizon image. The heat-dashboard was recently packaged in RDO [1]
but hasn't yet made it's way into current-tripleo or current-passed-ci.

We need to add the package into the kolla image once there is a RDO
promotion and re-enable the plugin.

[1] https://github.com/rdo-packages/heat-dashboard-distgit

Change-Id: I7c0ccd915d8429bb348c09ab621163aef275c669
Related-Bug: #1752132